[BREAKING COMMIT] Passing xk to solve leads to error
In test_rgm.jl I try to pass xk (which is returned by solve) again
to solve in an iterative process.

If I do this, line 555 of ExaPF.jl throws an error. This seems like
an issue with ConvergenceStatus.

I checked that the types of x and xk where the same. Not sure what
is causing the issue.

In the mean-time I am commenting out "convergencestatus" and returning
a dummy value so I can proceed with RGM.